+++ src/types/PaymentSheet.ts
@@ -360,6 +360,10 @@ export enum RowStyle {
   FloatingButton = 'floatingButton',
   /** A flat style with a checkmark */
   FlatWithCheckmark = 'flatWithCheckmark',
+  /** A flat style with a chevron
+   * Note that the EmbeddedPaymentElementConfiguration.rowSelectionBehavior must be set to `immediateAction` to use this style.
+   */
+  FlatWithChevron = 'flatWithChevron',
 }
 
 /** Describes the appearance of the radio button */
@@ -383,6 +387,14 @@ export interface CheckmarkConfig {
   color?: ThemedColor;
 }
 
+/** Describes the appearance of the chevron */
+export interface ChevronConfig {
+  /** The color of the chevron, represented as a hex string #AARRGGBB or #RRGGBB.
+   * @default The iOS or Android system gray color
+   */
+  color?: ThemedColor;
+}
+
 /** Describes the appearance of the flat style row */
 export interface FlatConfig {
   /** The thickness of the separator line between rows.
@@ -397,7 +409,7 @@ export interface FlatConfig {
 
   /** The insets of the separator line between rows.
    * @default { top: 0, left: 30, bottom: 0, right: 0 } for RowStyle.FlatWithRadio
-   * @default { top: 0, left: 0, bottom: 0, right: 0 } for RowStyle.FlatWithCheckmark and RowStyle.FloatingButton
+   * @default { top: 0, left: 0, bottom: 0, right: 0 } for RowStyle.FlatWithCheckmark, RowStyle.FlatWithChevron, and RowStyle.FloatingButton
    */
   separatorInsets?: EdgeInsetsConfig;
 
@@ -416,6 +428,9 @@ export interface FlatConfig {
 
   /** Appearance settings for the checkmark (used when RowStyle is FlatWithCheckmark) */
   checkmark?: CheckmarkConfig;
+
+  /** Appearance settings for the chevron (used when RowStyle is FlatWithChevron) */
+  chevron?: ChevronConfig;
 }
 
 /** Describes the appearance of the floating button style payment method row */
@@ -564,6 +579,17 @@ export type PaymentMode = {
   /* Controls when the funds will be captured.
   Seealso: https://stripe.com/docs/api/payment_intents/create#create_payment_intent-capture_method */
   captureMethod?: CaptureMethod;
+  /** Additional payment method options params.
+  Seealso: https://docs.stripe.com/api/payment_intents/create#create_payment_intent-payment_method_options */
+  paymentMethodOptions?: PaymentMethodOptions;
+};
+
+export type PaymentMethodOptions = {
+  /* This is an experimental feature that may be removed at any time
+  A map of payment method types to setup_future_usage value. (e.g. card: 'OffSession') */
+  setupFutureUsageValues: {
+    [key: string]: FutureUsage;
+  };
 };
 
 /* Use this if your integration creates a SetupIntent */