+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+**Features**
+- Added rowSelectionBehavior to `EmbeddedPaymentElementConfiguration` with `immediateAction` option
+- Added `flatWithChevron` to `AppearanceParams.embeddedPaymentElement.rowConfig.style`
+- - Added `PaymentMethodOptions` to `PaymentMode` to enable setting payment method level setup future usage value
+- Added `None` to `FutureUsage`
+
 ## 0.48.0 - 2025-06-11
 
 **Feature**